feat. 리뷰 반영: score 계산 로직 winningTicket -> tickets
