+++ src/main/kotlin/lotto/domain/WinningTicket.kt
@@ -7,8 +7,4 @@ data class WinningTicket(val winningLotto: Lotto, val bonus: Int) {
     init {
         require(!winningLotto.contains(bonus)) { BONUS_EXCEPTION }
     }
-
-    fun score(tickets: Tickets): Score {
-        return Score(tickets.tickets.map { Rank.getValue(this.winningLotto.getSameCount(it), it.contains(this.bonus)) })
-    }
 }
