+++ src/test/kotlin/lotto/domain/LotteryShopTest.kt
@@ -3,8 +3,6 @@ package lotto.domain
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.CsvSource
 
 internal class LotteryShopTest {
     @Test
